Make BasicUI available to external packages
The BasicUI needs to implement the terminal.UI, and does not
support coloring.
